requirements.txt
attrs==21.4.0
bidict==0.22.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.3.1
huggingface-hub==0.4.0
hypothesis==6.43.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
#kenlm==0.0.0
MarkupSafe==2.1.0
numpy==1.22.3
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pycparser==2.21
pyctcdecode==0.1.0
pygtrie==2.4.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
six==1.16.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
tokenizers==0.11.6
torch==1.11.0
torchaudio==0.11.0
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.1.1
urllib3==1.26.8
webrtcvad==2.0.10
Werkzeug==2.0.3
openai-whisper
